Rename interceptor binders to interceptor adapters.
